refactor: 모임 수요 댓글 서비스 로직 분리
